+++ provider/paymentmethod/datasource_payment_method_test.go
@@ -1,44 +1,21 @@
 package paymentmethod_test
 
 import (
-	"os"
 	"regexp"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 
-	"github.com/RedisLabs/terraform-provider-rediscloud/provider"
+	"github.com/RedisLabs/terraform-provider-rediscloud/provider/testhelpers"
 	"github.com/RedisLabs/terraform-provider-rediscloud/provider/utils"
 )
-
-var protoV5ProviderFactories = map[string]func() (tfprotov5.ProviderServer, error){
-	"rediscloud": func() (tfprotov5.ProviderServer, error) {
-		muxServer, err := provider.MuxProviderServerCreator(
-			provider.NewSdkProvider("dev")(),
-			provider.NewFrameworkProvider("dev")(),
-		)
-		if err != nil {
-			return nil, err
-		}
-		return muxServer(), nil
-	},
-}
-
-func testAccPreCheck(t *testing.T) {
-	for _, name := range []string{"REDISCLOUD_URL", "REDISCLOUD_ACCESS_KEY", "REDISCLOUD_SECRET_KEY"} {
-		if _, ok := os.LookupEnv(name); !ok {
-			t.Fatalf("Missing `%s` environment variable", name)
-		}
-	}
-}
 
 func TestAccDataSourceRedisCloudPaymentMethod_basic(t *testing.T) {
 	utils.AccRequiresEnvVar(t, "EXECUTE_TESTS")
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV5ProviderFactories: protoV5ProviderFactories,
+		PreCheck:                 func() { testhelpers.BasicPreCheck(t) },
+		ProtoV5ProviderFactories: testhelpers.ProtoV5ProviderFactories(),
 		CheckDestroy:             nil, // payment method isn't managed by this provider
 		Steps: []resource.TestStep{
 			{