+++ provider/pro/datasource_rediscloud_pro_subscription.go
@@ -146,6 +146,12 @@ func DataSourceRedisCloudProSubscription() *schema.Resource {
 								},
 							},
 						},
+						"resource_tags": {
+							Description: "A map of resource tags associated with this subscription.",
+							Type:        schema.TypeMap,
+							Computed:    true,
+							Elem:        &schema.Schema{Type: schema.TypeString},
+						},
 					},
 				},
 			},