+++ provider/regions/testdata/datasource_filtered.tf
@@ -1,0 +1,7 @@
+variable "provider_name" {
+  type = string
+}
+
+data "rediscloud_regions" "example" {
+  provider_name = var.provider_name
+}